+++ cloudmanager/data_source_netapp_cloudmanager_aws_fsx.go
@@ -18,7 +18,6 @@ func dataSourceAWSFSX() *schema.Resource {
 			"id": {
 				Type:     schema.TypeString,
 				Required: true,
-				ForceNew: true,
 			},
 			"status": {
 				Type:     schema.TypeString,